+++ Python_BackEnd/Products/Models/SpecificProduct.py
@@ -1,5 +1,6 @@
 import uuid
 from dataclasses import dataclass
+from decimal import Decimal
 
 
 @dataclass
@@ -9,6 +10,6 @@ class SpecificProduct:
     description: str = ""
     imagePath: str = ""
     retailer: str = ""
-    cost: float = 0.0
+    cost: Decimal = 0.0
     link: str = ""
     product_type: str = ""
